# proxy/dist/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"DelegationAndRewards": {
				"properties": {
					"delegate": {
						"type": "string"
					},
					"rewards": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AccountDelegationAndRewardsRequestBody": {
				"properties": {
					"stakeKeyHashes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"stakeKeyHashes"
				],
				"type": "object"
			},
			"AdaHandle": {
				"properties": {
					"key": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"key",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Reward": {
				"properties": {
					"epoch": {
						"type": "number",
						"format": "double"
					},
					"amount": {
						"type": "string"
					},
					"pool_id": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"required": [
					"epoch",
					"amount",
					"pool_id",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddressAfter": {
				"properties": {
					"tx": {
						"type": "string"
					},
					"block": {
						"type": "string"
					}
				},
				"required": [
					"tx",
					"block"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddressUsedRequestBody": {
				"properties": {
					"untilBlock": {
						"type": "string"
					},
					"after": {
						"$ref": "#/components/schemas/AddressAfter"
					},
					"addresses": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"untilBlock",
					"after",
					"addresses"
				],
				"type": "object"
			},
			"Signature": {
				"properties": {
					"publicKey": {
						"type": "string"
					},
					"signature": {
						"type": "string"
					}
				},
				"required": [
					"publicKey",
					"signature"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Name": {
				"properties": {
					"signatures": {
						"items": {
							"$ref": "#/components/schemas/Signature"
						},
						"type": "array"
					},
					"sequenceNumber": {
						"type": "number",
						"format": "double"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"signatures",
					"sequenceNumber",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Url": {
				"properties": {
					"signatures": {
						"items": {
							"$ref": "#/components/schemas/Signature"
						},
						"type": "array"
					},
					"sequenceNumber": {
						"type": "number",
						"format": "double"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"signatures",
					"sequenceNumber",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Description": {
				"properties": {
					"signatures": {
						"items": {
							"$ref": "#/components/schemas/Signature"
						},
						"type": "array"
					},
					"sequenceNumber": {
						"type": "number",
						"format": "double"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"signatures",
					"sequenceNumber",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Logo": {
				"properties": {
					"signatures": {
						"items": {
							"$ref": "#/components/schemas/Signature"
						},
						"type": "array"
					},
					"sequenceNumber": {
						"type": "number",
						"format": "double"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"signatures",
					"sequenceNumber",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Ticker": {
				"properties": {
					"signatures": {
						"items": {
							"$ref": "#/components/schemas/Signature"
						},
						"type": "array"
					},
					"sequenceNumber": {
						"type": "number",
						"format": "double"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"signatures",
					"sequenceNumber",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Decimals": {
				"properties": {
					"signatures": {
						"items": {
							"$ref": "#/components/schemas/Signature"
						},
						"type": "array"
					},
					"sequenceNumber": {
						"type": "number",
						"format": "double"
					},
					"value": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"signatures",
					"sequenceNumber",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TokenInfo": {
				"properties": {
					"name": {
						"$ref": "#/components/schemas/Name"
					},
					"url": {
						"$ref": "#/components/schemas/Url"
					},
					"description": {
						"$ref": "#/components/schemas/Description"
					},
					"logo": {
						"$ref": "#/components/schemas/Logo"
					},
					"ticker": {
						"$ref": "#/components/schemas/Ticker"
					},
					"subject": {
						"type": "string"
					},
					"decimals": {
						"$ref": "#/components/schemas/Decimals"
					}
				},
				"required": [
					"name",
					"url",
					"description",
					"logo",
					"ticker",
					"subject",
					"decimals"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CIP25": {
				"properties": {
					"policyId": {
						"type": "string"
					},
					"assets": {
						"items": {
							"properties": {
								"metadata": {
									"type": "string"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"metadata",
								"name"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"policyId",
					"assets"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AssetName": {
				"type": "string"
			},
			"MetadataNftRequestBody": {
				"properties": {
					"assets": {
						"properties": {},
						"additionalProperties": {
							"items": {
								"$ref": "#/components/schemas/AssetName"
							},
							"type": "array"
						},
						"type": "object"
					}
				},
				"required": [
					"assets"
				],
				"type": "object"
			},
			"Block": {
				"properties": {
					"slot": {
						"type": "number",
						"format": "double"
					},
					"epoch": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"hash": {
						"type": "string"
					},
					"era": {
						"type": "number",
						"format": "double"
					},
					"isValid": {
						"type": "boolean"
					},
					"indexInBlock": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"slot",
					"epoch",
					"height",
					"hash",
					"era",
					"isValid",
					"indexInBlock"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pool": {
				"properties": {
					"pool_id": {
						"type": "string"
					},
					"hex": {
						"type": "string"
					},
					"vrf_key": {
						"type": "string"
					},
					"blocks_minted": {
						"type": "number",
						"format": "double"
					},
					"blocks_epoch": {
						"type": "number",
						"format": "double"
					},
					"live_stake": {
						"type": "string"
					},
					"live_size": {
						"type": "number",
						"format": "double"
					},
					"live_saturation": {
						"type": "number",
						"format": "double"
					},
					"live_delegators": {
						"type": "number",
						"format": "double"
					},
					"active_stake": {
						"type": "string"
					},
					"active_size": {
						"type": "number",
						"format": "double"
					},
					"declared_pledge": {
						"type": "string"
					},
					"live_pledge": {
						"type": "string"
					},
					"margin_cost": {
						"type": "number",
						"format": "double"
					},
					"fixed_cost": {
						"type": "string"
					},
					"reward_account": {
						"type": "string"
					},
					"owners": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"registration": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"retirement": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"pool_id",
					"hex",
					"vrf_key",
					"blocks_minted",
					"blocks_epoch",
					"live_stake",
					"live_size",
					"live_saturation",
					"live_delegators",
					"active_stake",
					"active_size",
					"declared_pledge",
					"live_pledge",
					"margin_cost",
					"fixed_cost",
					"reward_account",
					"owners",
					"registration",
					"retirement"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Transaction": {
				"properties": {
					"payload": {
						"type": "string"
					},
					"hash": {
						"type": "string"
					}
				},
				"required": [
					"payload",
					"hash"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionData": {
				"properties": {
					"transaction": {
						"$ref": "#/components/schemas/Transaction"
					},
					"block": {
						"$ref": "#/components/schemas/Block"
					}
				},
				"required": [
					"transaction",
					"block"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionHistoryRequestBody": {
				"properties": {
					"relationFilter": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"untilBlock": {
						"type": "string"
					},
					"after": {
						"$ref": "#/components/schemas/AddressAfter"
					},
					"addresses": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"untilBlock",
					"after",
					"addresses"
				],
				"type": "object"
			},
			"Utxo": {
				"properties": {
					"index": {
						"type": "number",
						"format": "double"
					},
					"txHash": {
						"type": "string"
					},
					"payload": {
						"type": "string"
					}
				},
				"required": [
					"index",
					"txHash",
					"payload"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UtxoData": {
				"properties": {
					"utxo": {
						"$ref": "#/components/schemas/Utxo"
					},
					"block": {
						"$ref": "#/components/schemas/Block"
					}
				},
				"required": [
					"utxo",
					"block"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UtxoPointers": {
				"properties": {
					"index": {
						"type": "number",
						"format": "double"
					},
					"txHash": {
						"type": "string"
					}
				},
				"required": [
					"index",
					"txHash"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionOutputRequestBody": {
				"properties": {
					"utxoPointers": {
						"items": {
							"$ref": "#/components/schemas/UtxoPointers"
						},
						"type": "array"
					}
				},
				"required": [
					"utxoPointers"
				],
				"type": "object"
			},
			"TransactionSubmitRequestBody": {
				"properties": {
					"cbor": {
						"type": "string"
					}
				},
				"required": [
					"cbor"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "OneBox API",
		"version": "1.0.0",
		"description": "API specification",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/account/state": {
			"post": {
				"operationId": "GetDelegationAndRewards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"additionalProperties": {
										"$ref": "#/components/schemas/DelegationAndRewards"
									},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Account"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AccountDelegationAndRewardsRequestBody"
							}
						}
					}
				}
			}
		},
		"/address/handle/{handle}": {
			"get": {
				"operationId": "GetAddressForHandler",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AdaHandle"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Address"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "handle",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/address/rewards/{stakeAddress}": {
			"get": {
				"operationId": "GetRewards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Reward"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Address"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "stakeAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/address/used": {
			"post": {
				"operationId": "GetAddressUsed",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"type": "string"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Address"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddressUsedRequestBody"
							}
						}
					}
				}
			}
		},
		"/assets/metadata/{policyId}/{assetName}": {
			"get": {
				"operationId": "GetMetadata",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenInfo"
								}
							}
						}
					}
				},
				"tags": [
					"Assets"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "policyId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "assetName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/assets/metadata/nft": {
			"post": {
				"operationId": "GetMetadataNft",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CIP25"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Assets"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MetadataNftRequestBody"
							}
						}
					}
				}
			}
		},
		"/block/latest": {
			"get": {
				"operationId": "GetLatest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Block"
								}
							}
						}
					}
				},
				"tags": [
					"Block"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/pool/{poolId}": {
			"get": {
				"operationId": "GetPoolInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Pool"
								}
							}
						}
					}
				},
				"tags": [
					"Pool"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "poolId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/pool": {
			"get": {
				"operationId": "GetPools",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Pool"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Pool"
				],
				"security": [],
				"parameters": []
			}
		},
		"/transaction/history": {
			"post": {
				"operationId": "GetTransactionHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TransactionData"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Transaction"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TransactionHistoryRequestBody"
							}
						}
					}
				}
			}
		},
		"/transaction/output": {
			"post": {
				"operationId": "GetTransactionOutput",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UtxoData"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Transaction"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TransactionOutputRequestBody"
							}
						}
					}
				}
			}
		},
		"/transaction/submit": {
			"post": {
				"operationId": "TransactionSubmit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Transaction"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TransactionSubmitRequestBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8000",
			"description": "Local development"
		}
	]
}